ci: add Unity-free Core lib + GitHub Actions test workflow
The xUnit tests compiled against UnityEngine and ONI's Assembly-CSharp (through the mod
project), so they could not build on a hosted runner with no game install. Extract the
genuinely game-independent logic into a new netstandard2.0 OniExtract2024.Core:

- UiImageRect (moved from the mod) — the uiImageRect crop->cell geometry.
- ExportPaths.BuildExportPath (lifted out of BaseExport, which now delegates to it).

Add OniExtract2024.Core.Tests (net8.0, xUnit) covering both, and a CI workflow that runs
just `dotnet test` on it — green on ubuntu-latest with no DLLs. The mod and the
game-coupled OniExtract2024.Tests (ImageCrop, converter, model POCOs that need Unity types)
stay a local Visual Studio build, out of CI by design.
