+++ OniExtract2024.Core/UiImageRect.cs
