+++ OniExtract2024.sln
@@ -7,6 +7,10 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OniExtract2024", "OniExtrac
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OniExtract2024.Tests", "OniExtract2024.Tests\OniExtract2024.Tests.csproj", "{58318822-3919-499E-938D-92392131C576}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OniExtract2024.Core", "OniExtract2024.Core\OniExtract2024.Core.csproj", "{1E1D2C3B-0A09-4F8E-9D6C-0000C0DE0001}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OniExtract2024.Core.Tests", "OniExtract2024.Core.Tests\OniExtract2024.Core.Tests.csproj", "{1E1D2C3B-0A09-4F8E-9D6C-0000C0DE0002}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -21,6 +25,14 @@ Global
 		{58318822-3919-499E-938D-92392131C576}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{58318822-3919-499E-938D-92392131C576}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{58318822-3919-499E-938D-92392131C576}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1E1D2C3B-0A09-4F8E-9D6C-0000C0DE0001}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1E1D2C3B-0A09-4F8E-9D6C-0000C0DE0001}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1E1D2C3B-0A09-4F8E-9D6C-0000C0DE0001}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1E1D2C3B-0A09-4F8E-9D6C-0000C0DE0001}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1E1D2C3B-0A09-4F8E-9D6C-0000C0DE0002}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1E1D2C3B-0A09-4F8E-9D6C-0000C0DE0002}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1E1D2C3B-0A09-4F8E-9D6C-0000C0DE0002}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1E1D2C3B-0A09-4F8E-9D6C-0000C0DE0002}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
